Add session management with redis
Добавлено хранение сессий и авторизация по ним

+++ Backend/src/main/java/com/messenger/prism/repository/ChatRepo.java
@@ -3,5 +3,5 @@ package com.messenger.prism.repository;
 import com.messenger.prism.entity.ChatEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ChatRepo extends CrudRepository<ChatEntity, Long> {
+public interface ChatRepo extends CrudRepository<ChatEntity, Integer> {
 }
